+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-### 2.4.5 (TBD)
+### 2.4.5 (October 15, 2021)
 
 - Feature: Intercepting headless services is now supported. It's now possible to request a headless service on whatever port it exposes and get a response from the intercept.